Move `Block.SlotNo` to `Cardano.Read.Ledger.…`
