+++ lib/read/lib/Cardano/Read/Ledger/Block/SlotNo.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Cardano.Wallet.Read.Block.SlotNo
+module Cardano.Read.Ledger.Block.SlotNo
     ( getEraSlotNo
     , SlotNo (..)
     , prettySlotNo
