+++ lib/read/lib/Cardano/Wallet/Read/Block/Gen/Babbage.hs
@@ -37,6 +37,9 @@ import Cardano.Protocol.TPraos.BHeader
 import Cardano.Read.Ledger.Block.BlockNo
     ( BlockNo (..)
     )
+import Cardano.Read.Ledger.Block.SlotNo
+    ( SlotNo (..)
+    )
 import Cardano.Wallet.Read
     ( TxT
     )
@@ -52,9 +55,6 @@ import Cardano.Wallet.Read.Block.Gen.Shelley
     , mkKeyVRF'
     , mkSignedKES
     , oCertamente
-    )
-import Cardano.Wallet.Read.Block.SlotNo
-    ( SlotNo (..)
     )
 import Ouroboros.Consensus.Protocol.Praos
     ( Praos
