+++ lib/read/lib/Cardano/Wallet/Read/Block/Gen/Byron.hs
@@ -28,11 +28,11 @@ import Cardano.Crypto.Signing
 import Cardano.Read.Ledger.Block.BlockNo
     ( BlockNo (..)
     )
+import Cardano.Read.Ledger.Block.SlotNo
+    ( SlotNo (..)
+    )
 import Cardano.Wallet.Read.Block.Gen.BlockParameters
     ( BlockParameters (..)
-    )
-import Cardano.Wallet.Read.Block.SlotNo
-    ( SlotNo (..)
     )
 import Cardano.Wallet.Read.Eras
     ( Byron
